requirements.txt
# Enhanced Underwriting RAG System - Requirements

# ============================================================================
# Streamlit
# ============================================================================
streamlit>=1.31.0

# ============================================================================
# 文档处理
# ============================================================================
PyMuPDF>=1.23.0              # PDF处理
python-docx>=0.8.11          # DOCX处理
pytesseract>=0.3.10          # OCR识别
Pillow>=10.0.0               # 图像处理

# ============================================================================
# 向量存储和检索
# ============================================================================
faiss-cpu>=1.7.4             # FAISS向量检索（CPU版本）
# 如果有GPU，使用: faiss-gpu>=1.7.4

# ============================================================================
# Embeddings和AI
# ============================================================================
openai>=1.0.0                # OpenAI API（用于Embeddings和LLM）
# 可选：本地Embedding模型
# sentence-transformers>=2.2.2

# ============================================================================
# 基础工具库
# ============================================================================
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# ============================================================================
# 可选：数据库和存储
# ============================================================================
# chromadb>=0.4.0            # 替代向量数据库
# pinecone-client>=2.2.0     # 云端向量数据库

# ============================================================================
# 开发和测试
# ============================================================================
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
